backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
langchain==0.1.9
langchain-anthropic==0.1.4
anthropic==0.18.1
boto3==1.34.47
psycopg2-binary==2.9.9
qdrant-client==1.7.3
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
uuid==1.30
celery[redis]==5.3.6
flower==2.0.1
pytest==7.4.0
redis==5.0.1
celery-redbeat==2.0.0
